Comparison should be made between primatives.
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@19590 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/event/UserDatasetEventListHandler.java
@@ -65,7 +65,7 @@ public class UserDatasetEventListHandler extends BaseCLI {
     for (UserDatasetEvent event : eventList) {
     	
 
-      if (event.getEventId() <= lastHandledEventId) continue;
+      if ((long)event.getEventId() <= (long)lastHandledEventId) continue;
       
       if (event instanceof UserDatasetInstallEvent) {
         UserDatasetTypeHandler typeHandler = typeHandlers.get(event.getUserDatasetType());
